+++ freemarker-generator-tools/src/main/java/org/apache/freemarker/generator/tools/commonscsv/CommonsCSVTool.java
@@ -18,7 +18,6 @@ package org.apache.freemarker.generator.tools.commonscsv;
 
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
-import org.apache.commons.csv.CSVPrinter;
 import org.apache.commons.csv.CSVRecord;
 import org.apache.commons.io.input.BOMInputStream;
 import org.apache.freemarker.generator.base.datasource.DataSource;
@@ -26,7 +25,6 @@ import org.apache.freemarker.generator.base.util.StringUtils;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.Writer;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -37,7 +35,6 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static java.lang.Boolean.parseBoolean;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
@@ -45,10 +42,8 @@ import static org.apache.commons.io.IOUtils.toInputStream;
 
 public class CommonsCSVTool {
 
-    private final CSVFormat defaulCSVFormat = csvFormat();
-
     public CSVParser parse(DataSource dataSource) {
-        return parse(dataSource, defaulCSVFormat);
+        return parse(dataSource, defaultCSVInputFormat());
     }
 
     public CSVParser parse(DataSource dataSource, CSVFormat format) {
@@ -68,7 +63,7 @@ public class CommonsCSVTool {
     }
 
     public CSVParser parse(String csv) {
-        return parse(csv, defaulCSVFormat);
+        return parse(csv, defaultCSVInputFormat());
     }
 
     public CSVParser parse(String csv, CSVFormat format) {
@@ -88,17 +83,12 @@ public class CommonsCSVTool {
         return createCSVFormats();
     }
 
-    /**
-     * Get a CSVPrinter using the FreeMarker's writer instance.
-     *
-     * @param csvFormat CSV format to use for writing records
-     * @param writer    Writer to receive the CSV output
-     * @return CSVPrinter instance
-     * @throws IOException thrown if the parameters of the format are inconsistent or if either out or format are null.
-     */
-    public CSVPrinter printer(CSVFormat csvFormat, Writer writer) throws IOException {
-        // We do not close the CSVPrinter but the underlying writer at the of processing
-        return new CSVPrinter(writer, csvFormat);
+    public CommonsCSVPrinterFacade printer() throws IOException {
+        return printer(defaultCSVOutputFormat());
+    }
+
+    public CommonsCSVPrinterFacade printer(CSVFormat csvFormat) throws IOException {
+        return new CommonsCSVPrinterFacade(csvFormat);
     }
 
     /**
@@ -248,6 +238,7 @@ public class CommonsCSVTool {
     private static Map<String, CSVFormat> createCSVFormats() {
         final Map<String, CSVFormat> result = new HashMap<>();
         result.put("DEFAULT", CSVFormat.DEFAULT);
+        result.put("DATAFRAME", CSVFormat.DEFAULT.withDelimiter(';').withFirstRecordAsHeader());
         result.put("EXCEL", CSVFormat.EXCEL);
         result.put("INFORMIX_UNLOAD", CSVFormat.INFORMIX_UNLOAD);
         result.put("INFORMIX_UNLOAD_CSV", CSVFormat.INFORMIX_UNLOAD_CSV);
@@ -262,21 +253,22 @@ public class CommonsCSVTool {
         return result;
     }
 
-    private CSVFormat csvFormat() {
+    /**
+     * Provides a CSV default input format.
+     *
+     * @return CSV format
+     */
+    private static CSVFormat defaultCSVInputFormat() {
+        return CSVFormat.DEFAULT;
+    }
 
-        CSVFormat csvFormat = CSVFormat.valueOf(System.getProperty("CSV_TOOL_FORMAT", "Default"));
-
-        final String delimiter = System.getProperty("CSV_TOOL_DELIMITER");
-        if (StringUtils.isNotEmpty(delimiter)) {
-            csvFormat = csvFormat.withDelimiter(toDelimiter(delimiter));
-        }
-
-        final boolean withHeader = parseBoolean(System.getProperty("CSV_TOOL_HEADERS", Boolean.toString(!csvFormat.getSkipHeaderRecord())));
-        if (withHeader) {
-            csvFormat = csvFormat.withHeader();
-        }
-
-        return csvFormat;
+    /**
+     * Provides a CSV default output format.
+     *
+     * @return CSV format
+     */
+    private static CSVFormat defaultCSVOutputFormat() {
+        return CSVFormat.DEFAULT;
     }
 
     private static final class ValueResolver implements Function<CSVRecord, String> {