+++ travis.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 mvn clean install
 cd ./freemarker-generator-cli
-sh ./run-samples.sh
+sh ./run-examples.sh
 cd ../freemarker-generator-maven-plugin-sample
 mvn clean package